IMAGE TRANSITIONS FINALLY
using picasso
finally got transitions to work
after what may have been 20 programmer hours of logical and mental crisis, turns out the large set text on an unrelated text field in the fragment, it caused some unexplained, stacktrace-less OpenGL error when it tried to animate the transition.
when it didnt cause the app to crash, the closest it got was it animating the picture flying off screen.
finally commenting out the set text for that one big textfield, it just magically started working exactly as i thought it should have.
WELP
will figure out how to fill in the e-book text again later